rename getCurrentChar.js into currentHandler.js
